# perf: optimize hot paths based on assembly analysis