## Summary

This PR applies targeted micro-optimizations identified through assembly code analysis using `cargo-asm`. The optimizations focus on improving branch prediction and reducing redundant memory operations in hot paths.

## Changes

### 1. State Enum Explicit Values
- Set `Top = 0` for better branch prediction
- Allows compiler to use `cbz` (compare-branch-zero) instead of general comparisons
- First state check becomes `cbz w8, LBB1_18` in assembly (single instruction)

### 2. Cold Attribute on Error Path
- Mark `in_comment()` as `#[cold]`
- Improves instruction cache utilization for hot paths
- Error handling is moved out of the main code path

### 3. Byte Caching
- Cache character before `top()` modifies it in comma-checking logic
- Reduces redundant memory loads in the critical path
- `let byte = *c;` before calling `top(c)`

### 4. Inline Hint
- Add `#[inline]` to `consume_comment_whitespace_until_maybe_bracket`
- Helps compiler inline this frequently-called function

### 5. Consistent Match Dereferencing
- Use `*state` consistently in match arms
- Cleaner code and better compiler optimization opportunities

## Performance Impact

Benchmark results show mixed but generally neutral impact (within measurement noise):

| Benchmark | Baseline | Optimized | Change |
|-----------|----------|-----------|--------|
| tsconfig | 2.34µs | 2.44µs | +4.1% |
| no_comments | 504ns | 496ns | **-1.6%** ✓ |
| minimal_comments | 130ns | 132ns | +1.5% |
| large_with_comments | 12.29µs | 12.45µs | +1.3% |

The variations are within typical benchmark noise levels. The `no_comments` fast path shows a small improvement.

## Assembly Analysis

The compiler already does an excellent job optimizing this code:
- NEON SIMD instructions are used effectively for `memchr` operations
- State machine branches are well-predicted
- Memory bandwidth (for `fill` operations) is the primary bottleneck

These optimizations provide better hints to the compiler and improve code clarity without sacrificing correctness.

## Testing

- ✅ All 35 unit tests pass
- ✅ All 3 doc tests pass
- ✅ Benchmarks run successfully

🤖 Generated with [Claude Code](https://claude.com/claude-code)